Make FA3 paged attention ready for upgrade to Cutlass 3.6
